Support floating point numbers in selector predicates

+++ crates/moqtail-core/src/ast.rs
@@ -27,21 +27,21 @@ pub enum Operator {
     Ge,
 }
 
-#[derive(Debug, PartialEq, Eq, Clone)]
+#[derive(Debug, PartialEq, Clone, Copy)]
 pub enum Value {
-    Number(i64),
+    Number(f64),
     Bool(bool),
     Str(String),
 }
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Debug, PartialEq)]
 pub struct Predicate {
     pub field: Field,
     pub op: Operator,
     pub value: Value,
 }
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Debug, PartialEq)]
 pub struct Step {
     pub axis: Axis,
     pub segment: Segment,
@@ -56,7 +56,7 @@ pub enum Stage {
     Count,
 }
 
-#[derive(Debug, PartialEq, Eq)]
+#[derive(Debug, PartialEq)]
 pub struct Selector {
     pub steps: Vec<Step>,
     pub stages: Vec<Stage>,
